% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hilo.R
\name{darken_fill}
\alias{darken_fill}
\title{darken_fill}
\usage{
darken_fill(col, prob)
}
\arguments{
\item{col}{colors}

\item{prob}{probability values}
}
\description{
darken fill colors for probability ranges
}
