% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{density.distribution}
\alias{density.distribution}
\title{The probability density/mass function}
\usage{
\method{density}{distribution}(x, at, ..., log = FALSE)
}
\arguments{
\item{x}{The distribution(s).}

\item{at}{The point at which to compute the density/mass.}

\item{...}{Additional arguments passed to methods.}

\item{log}{If \code{TRUE}, probabilities will be given as log probabilities.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Computes the probability density function for a continuous distribution, or
the probability mass function for a discrete distribution.
}
\seealso{
\code{vignette("broadcasting", package = "distributional")} for how the
\code{at} argument is combined with the distributions.
}
