% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_bernoulli.R
\name{dist_bernoulli}
\alias{dist_bernoulli}
\title{The Bernoulli distribution}
\usage{
dist_bernoulli(prob)
}
\arguments{
\item{prob}{The probability of success on each trial.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_bernoulli(prob = c(0.05, 0.5, 0.3, 0.9, 0.1))

}
