% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_cauchy.R
\name{dist_cauchy}
\alias{dist_cauchy}
\title{The Cauchy distribution}
\usage{
dist_cauchy(location, scale)
}
\arguments{
\item{location}{location and scale parameters.}

\item{scale}{location and scale parameters.}
}
\description{
\lifecycle{maturing}
}
\examples{
dist_cauchy(location = c(0, 0, 0, -2), scale = c(0.5, 1, 2, 1))

}
\seealso{
\link[stats:Cauchy]{stats::Cauchy}
}
