% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degenerate.R
\name{dist_degenerate}
\alias{dist_degenerate}
\title{The degenerate distribution}
\usage{
dist_degenerate(x)
}
\arguments{
\item{x}{The value of the distribution.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_degenerate(x = 1:5)

}
