% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_f.R
\name{dist_f}
\alias{dist_f}
\title{The (non-central) Chi-Squared Distribution}
\usage{
dist_f(df1, df2, ncp = NULL)
}
\arguments{
\item{ncp}{non-centrality parameter (non-negative).}
}
\description{
\lifecycle{stable}
}
\examples{
dist_f(df1 = c(1,2,5,10,100), df2 = c(1,1,2,1,100))

}
\seealso{
df
}
