% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_geometric.R
\name{dist_geometric}
\alias{dist_geometric}
\title{The Geometric Distribution}
\usage{
dist_geometric(prob)
}
\arguments{
\item{prob}{probability of success in each trial. \code{0 < prob <= 1}.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_geometric(prob = c(0.2, 0.5, 0.8))

}
\seealso{
\link[stats:Geometric]{stats::Geometric}
}
