% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_gumbel.R
\name{dist_gumbel}
\alias{dist_gumbel}
\title{The Gumbel distribution}
\usage{
dist_gumbel(alpha, scale)
}
\arguments{
\item{alpha}{location parameter.}

\item{scale}{parameter. Must be strictly positive.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_gumbel(alpha = c(0.5, 1, 1.5, 3), scale = c(2, 2, 3, 4))

}
\seealso{
\link[actuar:Gumbel]{actuar::Gumbel}
}
