% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_inverse_exponential.R
\name{dist_inverse_exponential}
\alias{dist_inverse_exponential}
\title{The Inverse Exponential distribution}
\usage{
dist_inverse_exponential(rate)
}
\arguments{
\item{rate}{an alternative way to specify the scale.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_inverse_exponential(rate = 1:5)

}
\seealso{
\link[actuar:InverseExponential]{actuar::InverseExponential}
}
