% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_inverse_pareto.R
\name{dist_inverse_pareto}
\alias{dist_inverse_pareto}
\title{The Inverse Pareto Distribution}
\usage{
dist_inverse_pareto(shape, scale)
}
\arguments{
\item{shape, scale}{parameters. Must be strictly positive.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The Inverse Pareto distribution is often used to model heavy-tailed data.
}
\details{
We recommend reading this documentation on
\url{https://pkg.mitchelloharawild.com/distributional/}, where the math
will render nicely.

In the following, let \eqn{X} be an Inverse Pareto random variable with parameters
\code{shape} (\eqn{\alpha}) and \code{scale} (\eqn{\theta}).

\strong{Support}: \eqn{(0, \infty)}

\strong{Mean}: \eqn{\frac{\alpha \theta}{\alpha - 1}} for \eqn{\alpha > 1}

\strong{Variance}: \eqn{\frac{\alpha \theta^2}{(\alpha - 1)^2 (\alpha - 2)}} for \eqn{\alpha > 2}

\strong{Probability density function (p.d.f)}:

\deqn{
    f(x) = \frac{\alpha \theta^\alpha}{x^{\alpha + 1}}
  }{
    f(x) = \alpha \theta^\alpha / x^{\alpha + 1}
  }

\strong{Cumulative distribution function (c.d.f)}:

\deqn{
    F(x) = 1 - \left( \frac{\theta}{x} \right)^\alpha
  }{
    F(x) = 1 - (\theta / x)^\alpha
  }
}
\examples{
dist <- dist_inverse_pareto(shape = 2.5, scale = 3)
dist

\dontshow{if (requireNamespace("actuar", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mean(dist)
variance(dist)
generate(dist, 10)

density(dist, 2)
density(dist, 2, log = TRUE)

cdf(dist, 4)

quantile(dist, 0.7)
\dontshow{\}) # examplesIf}
}
\seealso{
\link[actuar:InversePareto]{actuar::dinvpareto}
}
