% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_logistic.R
\name{dist_logistic}
\alias{dist_logistic}
\title{The Logistic distribution}
\usage{
dist_logistic(location, scale)
}
\arguments{
\item{location}{location and scale parameters.}

\item{scale}{location and scale parameters.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_logistic(location = c(5,9,9,6,2), scale = c(2,3,4,2,1))

}
\seealso{
\link[stats:Logistic]{stats::Logistic}
}
