% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_pareto.R
\name{dist_pareto}
\alias{dist_pareto}
\title{The Pareto distribution}
\usage{
dist_pareto(shape, scale)
}
\arguments{
\item{shape}{parameters. Must be strictly positive.}

\item{scale}{parameters. Must be strictly positive.}
}
\description{
\lifecycle{questioning}
}
\examples{
dist_pareto(shape = c(10, 3, 2, 1), scale = rep(1, 4))

}
\seealso{
\link[actuar:Pareto]{actuar::Pareto}
}
