% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_percentile.R
\name{dist_percentile}
\alias{dist_percentile}
\title{Percentile distribution}
\usage{
dist_percentile(x, percentile)
}
\arguments{
\item{x}{A list of values}

\item{percentile}{A list of percentiles}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
dist <- dist_normal()
percentiles <- seq(0.01, 0.99, by = 0.01)
x <- vapply(percentiles, quantile, double(1L), x = dist)
dist_percentile(list(x), list(percentiles*100))

}
