% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_poisson.R
\name{dist_poisson}
\alias{dist_poisson}
\title{The Poisson Distribution}
\usage{
dist_poisson(lambda)
}
\arguments{
\item{lambda}{vector of (non-negative) means.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_poisson(lambda = c(1, 4, 10))

}
\seealso{
\link[stats:Poisson]{stats::Poisson}
}
