% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_poisson_inverse_gaussian.R
\name{dist_poisson_inverse_gaussian}
\alias{dist_poisson_inverse_gaussian}
\title{The Poisson-Inverse Gaussian distribution}
\usage{
dist_poisson_inverse_gaussian(mean, shape)
}
\arguments{
\item{mean}{parameters. Must be strictly positive. Infinite
    values are supported.}

\item{shape}{parameters. Must be strictly positive. Infinite
    values are supported.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_poisson_inverse_gaussian(mean = c(1), shape = c(0.4))

}
\seealso{
actuar::dpoisinvgauss
}
