% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_student_t.R
\name{dist_student_t}
\alias{dist_student_t}
\title{The (non-central) location-scale Student t Distribution}
\usage{
dist_student_t(df, mu = 0, sigma = 1, ncp = NULL)
}
\arguments{
\item{df}{degrees of freedom (\eqn{> 0}, maybe non-integer).  \code{df
      = Inf} is allowed.}

\item{mu}{The location parameter of the distribution.
If \code{ncp == 0} (or \code{NULL}), this is the median.}

\item{sigma}{The scale parameter of the distribution.}

\item{ncp}{non-centrality parameter \eqn{\delta}{delta};
    currently except for \code{rt()}, only for \code{abs(ncp) <= 37.62}.
    If omitted, use the central t distribution.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_student_t(df = c(1,2,5), mu = c(0,1,2), sigma = c(1,2,3))

}
\seealso{
\link[stats:TDist]{stats::TDist}
}
