% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated.R
\name{dist_truncated}
\alias{dist_truncated}
\title{Truncate a distribution}
\usage{
dist_truncated(dist, lower = -Inf, upper = Inf)
}
\arguments{
\item{dist}{The distribution(s) to truncate.}

\item{lower, upper}{The range of values to keep from a distribution.}
}
\description{
Truncate a distribution
}
