% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_weibull.R
\name{dist_weibull}
\alias{dist_weibull}
\title{The Weibull distribution}
\usage{
dist_weibull(shape, scale)
}
\arguments{
\item{shape}{shape and scale parameters, the latter defaulting to 1.}

\item{scale}{shape and scale parameters, the latter defaulting to 1.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_weibull(shape = c(1,2,3,5,9,7.5,0.5), scale = c(0.5,0.5,0.5,1,2,1,1))

}
\seealso{
dweibull
}
