% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fget}
\alias{fget}
\title{Get the transform, inverse or derivative functions from a distribution}
\usage{
fget(x, what)
}
\arguments{
\item{x}{A distribution object}

\item{what}{A string or name indicating the function to extract}
}
\value{
A function or list of functions
}
\description{
Get the transform, inverse or derivative functions from a distribution
}
\examples{
d <- exp(dist_gamma(1,1))
fget(d, "transform")
fget(d, "inverse")
fget(d, "deriv")
}
