% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hilo.R
\name{geom_hilo_ribbon}
\alias{geom_hilo_ribbon}
\title{Ribbon plots for hilo intervals}
\usage{
geom_hilo_ribbon(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated in favour of the ggdist package and will removed
in a future release of this package. Consider using \code{\link[ggdist:stat_lineribbon]{ggdist::stat_lineribbon()}} or
\code{\link[ggdist:geom_lineribbon]{ggdist::geom_lineribbon()}} as an appropriate alternative.

\code{geom_hilo_ribbon()} displays the interval defined by a hilo object. The
luminance of the shaded area indicates its confidence level. The shade colour
can be controlled by the \code{fill} aesthetic, however the luminance will be
overwritten to represent the confidence level.
}
\examples{
dist <- dist_normal(1:3, 1:3)
library(ggplot2)
ggplot(
  data.frame(x = rep(1:3, 2), interval = c(hilo(dist, 80), hilo(dist, 95)))
) +
  geom_hilo_ribbon(aes(x = x, hilo = interval))

}
\seealso{
\code{\link[=geom_hilo_linerange]{geom_hilo_linerange()}} for discrete hilo intervals (vertical lines)
}
