% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{hdr.distribution}
\alias{hdr.distribution}
\title{Highest density regions of probability distributions}
\usage{
\method{hdr}{distribution}(x, size = 95, n = 512, ...)
}
\arguments{
\item{x}{The distribution(s).}

\item{size}{The size of the interval (between 0 and 100).}

\item{n}{The resolution used to estimate the distribution's density.}

\item{...}{Additional arguments used by methods.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

This function is highly experimental and will change in the future. In
particular, improved functionality for object classes and visualisation tools
will be added in a future release.

Computes minimally sized probability intervals highest density regions.
}
