% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdr.R
\name{hdr}
\alias{hdr}
\title{Compute highest density regions}
\usage{
hdr(x, ...)
}
\arguments{
\item{x}{Object to create hilo from.}

\item{...}{Additional arguments used by methods.}
}
\description{
Used to extract a specified prediction interval at a particular confidence
level from a distribution.
}
