% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{hilo.distribution}
\alias{hilo.distribution}
\title{Probability intervals of a probability distribution}
\usage{
\method{hilo}{distribution}(x, size = 95, ...)
}
\arguments{
\item{x}{The distribution(s).}

\item{size}{The size of the interval (between 0 and 100).}

\item{...}{Additional arguments used by methods.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Returns a \code{hilo} central probability interval with probability coverage of
\code{size}. By default, the distribution's \code{\link[=quantile]{quantile()}} will be used to compute
the lower and upper bound for a centered interval
}
\seealso{
\code{\link[=hdr.distribution]{hdr.distribution()}}
}
