% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilo.R
\name{hilo}
\alias{hilo}
\title{Compute intervals}
\usage{
hilo(x, ...)
}
\arguments{
\item{x}{Object to create hilo from.}

\item{...}{Additional arguments used by methods.}
}
\description{
Used to extract a specified prediction interval at a particular confidence
level from a distribution.
}
