% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{is_distribution}
\alias{is_distribution}
\title{Test if the object is a distribution}
\usage{
is_distribution(x)
}
\arguments{
\item{x}{An object.}
}
\value{
TRUE if the object inherits from the distribution class.
}
\description{
This function returns \code{TRUE} for distributions and \code{FALSE} for all other objects.
\lifecycle{stable}
}
\examples{
# A distribution ----
dist <- dist_normal()
is_distribution(dist)

# distribution columns ----
df <- tibble(a = 1:10, b = dist_poisson(1:10), c = dist_normal(1:10))
df
mutate(df, across(is_distribution, mean))
}
