% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{kurtosis}
\alias{kurtosis}
\alias{kurtosis.distribution}
\title{Kurtosis of a probability distribution}
\usage{
kurtosis(x, ...)

\method{kurtosis}{distribution}(x, ...)
}
\arguments{
\item{x}{The distribution(s).}

\item{...}{Additional arguments used by methods.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
