% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{likelihood}
\alias{likelihood}
\alias{likelihood.distribution}
\title{The (log) likelihood of a sample matching a distribution}
\usage{
likelihood(x, ...)

\method{likelihood}{distribution}(x, sample, ..., log = FALSE)
}
\arguments{
\item{x}{The distribution(s).}

\item{...}{Additional arguments used by methods.}

\item{sample}{A list of sampled values to compare to distribution(s).}

\item{log}{If \code{TRUE}, the log-likelihood will be computed.}
}
\description{
\lifecycle{maturing}
}
