% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{median.distribution}
\alias{median.distribution}
\title{Median of a probability distribution}
\usage{
\method{median}{distribution}(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{The distribution(s).}

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}

\item{...}{Additional arguments used by methods.}
}
\description{
\lifecycle{stable}
}
\details{
Returns the median (50th percentile) of a probability distribution. This is
equivalent to \code{quantile(x, p=0.5)}.
}
