% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{new_dist}
\alias{new_dist}
\title{Create a new distribution}
\usage{
new_dist(..., class = NULL)
}
\arguments{
\item{...}{Parameters of the distribution (named)}

\item{class}{The class for S3 dispatch}
}
\description{
Create a new distribution
}
