% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdr.R
\name{new_hdr}
\alias{new_hdr}
\title{Construct hdr intervals}
\usage{
new_hdr(x = list())
}
\arguments{
\item{lower, upper}{A numeric vector of values for lower and upper limits.}

\item{size}{Size of the interval between [0, 1].}
}
\value{
A "hdr" vector
}
\description{
Construct hdr intervals
}
\examples{
new_hilo(lower = rnorm(10), upper = rnorm(10) + 5, size = 0.95)

}
\author{
Mitchell O'Hara-Wild
}
