% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilo.R
\name{new_hilo}
\alias{new_hilo}
\title{Construct hilo intervals}
\usage{
new_hilo(lower, upper, size)
}
\arguments{
\item{lower, upper}{A numeric vector of values for lower and upper limits.}

\item{size}{Size of the interval between \link{0, 1}.}
}
\value{
A "hilo" vector
}
\description{
Construct hilo intervals
}
\examples{
new_hilo(lower = rnorm(10), upper = rnorm(10) + 5, size = 0.95)

}
\author{
Earo Wang & Mitchell O'Hara-Wild
}
