% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{new_support_region}
\alias{new_support_region}
\title{Create a new support region vector}
\usage{
new_support_region(x, limits = NULL, interval = NULL)
}
\arguments{
\item{x}{A list of prototype vectors defining the distribution type.}

\item{limits}{A list of value limits for the distribution.}

\item{interval}{The}
}
\description{
Create a new support region vector
}
