% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{quantile.distribution}
\alias{quantile.distribution}
\title{Distribution Quantiles}
\usage{
\method{quantile}{distribution}(x, p, ..., log = FALSE)
}
\arguments{
\item{x}{The distribution(s).}

\item{p}{The probability of the quantile.}

\item{...}{Additional arguments passed to methods.}

\item{log}{If \code{TRUE}, probabilities will be given as log probabilities.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Computes the quantiles of a distribution.
}
\seealso{
\code{vignette("broadcasting", package = "distributional")} for how the
\code{p} argument is combined with the distributions.
}
