% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transformation-functions}
\alias{transformation-functions}
\alias{eval_transform}
\alias{eval_inverse}
\alias{eval_deriv}
\title{Evaluate the transformation, inverse or the jacobian of the inverse transformation of a random variable at a vector of values}
\usage{
eval_transform(dist, at)

eval_inverse(dist, at)

eval_deriv(dist, at)
}
\arguments{
\item{dist}{A distribution object}

\item{at}{A vector of values at which to evaluate the transformation, inverse or jacobian}
}
\value{
A vector or list of numeric values. If length(at) = 1, a numeric vector the same length as \code{dist} is returned. If length(at) > 1, a list of numeric vectors is returned, each list entry corresponding to a distribution in \code{dist}, and each vector the same length as \code{at}.
}
\description{
Evaluate the transformation, inverse or the jacobian of the inverse transformation of a random variable at a vector of values
}
