% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{variance}
\alias{variance}
\title{Distribution variance}
\usage{
variance(x, ...)
}
\arguments{
\item{x}{A distribution}

\item{...}{Additional arguments used by methods.}
}
\description{
Distribution variance
}
