% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R, R/hilo.R
\name{vec_arith.distribution}
\alias{vec_arith.distribution}
\alias{vec_ptype2.distribution}
\alias{vec_ptype2.distribution.default}
\alias{vec_ptype2.distribution.distribution}
\alias{vec_cast.distribution}
\alias{vec_cast.distribution.default}
\alias{vec_cast.distribution.distribution}
\alias{vec_math.hilo}
\alias{vec_arith.hilo}
\alias{vec_arith.numeric.hilo}
\title{vctrs compatibility functions}
\usage{
\method{vec_arith}{distribution}(op, x, y, ...)

vec_ptype2.distribution(x, y, ...)

\method{vec_ptype2}{distribution.default}(x, y, ..., x_arg = "x", y_arg = "y")

\method{vec_ptype2}{distribution.distribution}(x, y, ...)

vec_cast.distribution(x, to, ...)

\method{vec_cast}{distribution.default}(x, to, ...)

\method{vec_cast}{distribution.distribution}(x, to, ...)

\method{vec_math}{hilo}(.fn, .x, ...)

\method{vec_arith}{hilo}(op, x, y, ...)

\method{vec_arith}{numeric.hilo}(op, x, y, ...)
}
\description{
These methods allow distributional objects to work with vctrs.
}
