% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.distribution}
\alias{autoplot.distribution}
\title{Plot a distribution}
\usage{
autoplot.distribution(x, ...)
}
\arguments{
\item{x}{The distribution(s) to plot.}

\item{...}{Unused.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

This function is now defunct and can no longer be used. Instead consider using
the {ggdist} package to produce your own distribution plots. You can learn
more about how this plot can be produced using {ggdist} here:
https://mjskay.github.io/ggdist/articles/slabinterval.html
}
\keyword{internal}
