% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{cdf}
\alias{cdf}
\title{Cumulative distribution function}
\usage{
cdf(x, q, ...)
}
\arguments{
\item{x}{A distribution.}

\item{q}{The quantile at which the cdf is calculated.}
}
\description{
Cumulative distribution function
}
