% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{covariance}
\alias{covariance}
\title{Covariance}
\usage{
covariance(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments used by methods.}
}
\description{
A generic function for computing the covariance of an object.
}
\seealso{
\code{\link[=covariance.distribution]{covariance.distribution()}}, \code{\link[=variance]{variance()}}
}
