% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_beta.R
\name{dist_beta}
\alias{dist_beta}
\title{The Beta distribution}
\usage{
dist_beta(shape1, shape2)
}
\arguments{
\item{shape1, shape2}{The non-negative shape parameters of the Beta distribution.}
}
\description{
\lifecycle{maturing}
}
\examples{
dist_beta(shape1 = c(0.5, 5, 1, 2, 2), shape2 = c(0.5, 1, 3, 2, 5))

}
\seealso{
\link[stats:Beta]{stats::Beta}
}
