% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_binomial.R
\name{dist_binomial}
\alias{dist_binomial}
\title{The Binomial distribution}
\usage{
dist_binomial(size, prob)
}
\arguments{
\item{size}{The number of trials.}

\item{prob}{The probability of success on each trial.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_binomial(size = 1:5, prob = c(0.05, 0.5, 0.3, 0.9, 0.1))

}
