% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_burr.R
\name{dist_burr}
\alias{dist_burr}
\title{The Burr distribution}
\usage{
dist_burr(shape1, shape2, rate = 1)
}
\arguments{
\item{shape1}{parameters. Must be strictly positive.}

\item{shape2}{parameters. Must be strictly positive.}

\item{rate}{an alternative way to specify the scale.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_burr(shape1 = c(1,1,1,2,3,0.5), shape2 = c(1,2,3,1,1,2))

}
\seealso{
\link[actuar:Burr]{actuar::Burr}
}
