% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_chisq.R
\name{dist_chisq}
\alias{dist_chisq}
\title{The (non-central) Chi-Squared Distribution}
\usage{
dist_chisq(df, ncp = 0)
}
\arguments{
\item{df}{degrees of freedom (non-negative, but can be non-integer).}

\item{ncp}{non-centrality parameter (non-negative).}
}
\description{
\lifecycle{stable}
}
\examples{
dist_chisq(df = c(1,2,3,4,6,9))

}
\seealso{
\link[stats:Chisquare]{stats::Chisquare}
}
