% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_convolved.R
\name{dist_convolved}
\alias{dist_convolved}
\title{Convolve distributions}
\usage{
dist_convolved(x, y, n = 512, trunc = 1e-06)
}
\arguments{
\item{x, y}{A univariate distribution vector.}

\item{n}{the number of equally spaced points at which the convolution is to be estimated. This should be a power of two.}

\item{trunc}{The truncation of the support region if support of \code{x}, \code{y} are infinite.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{\link[=density]{density()}}, \code{\link[=mean]{mean()}}, and \code{\link[=variance]{variance()}} methods are approximate as
they are based on numerical derivatives.
}
\examples{

1

}
