% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_exponential.R
\name{dist_exponential}
\alias{dist_exponential}
\title{The Exponential Distribution}
\usage{
dist_exponential(rate)
}
\arguments{
\item{rate}{vector of rates.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_exponential(rate = c(2, 1, 2/3))

}
\seealso{
\link[stats:Exponential]{stats::Exponential}
}
