% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_f.R
\name{dist_f}
\alias{dist_f}
\title{The (non-central) Chi-Squared Distribution}
\usage{
dist_f(df1, df2, ncp = NULL)
}
\arguments{
\item{df1}{degrees of freedom.  \code{Inf} is allowed.}

\item{df2}{degrees of freedom.  \code{Inf} is allowed.}

\item{ncp}{non-centrality parameter. If omitted the central F is assumed.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_f(df1 = c(1,2,5,10,100), df2 = c(1,1,2,1,100))

}
\seealso{
\link[stats:FDist]{stats::FDist}
}
