% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_gamma.R
\name{dist_gamma}
\alias{dist_gamma}
\title{The Gamma distribution}
\usage{
dist_gamma(shape, rate)
}
\arguments{
\item{shape}{shape and scale parameters.  Must be positive,
    \code{scale} strictly.}

\item{rate}{an alternative way to specify the scale.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_gamma(shape = c(1,2,3,5,9,7.5,0.5), rate = c(0.5,0.5,0.5,1,2,1,1))

}
\seealso{
\link[stats:GammaDist]{stats::GammaDist}
}
