% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_hypergeometric.R
\name{dist_hypergeometric}
\alias{dist_hypergeometric}
\title{The Hypergeometric distribution}
\usage{
dist_hypergeometric(m, n, k)
}
\arguments{
\item{m}{The number of type I elements available.}

\item{n}{The number of type II elements available.}

\item{k}{The size of the sample taken.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_hypergeometric(m = rep(500, 3), n = c(50, 60, 70), k = c(100, 200, 300))

}
\seealso{
\link[stats:Hypergeometric]{stats::Hypergeometric}
}
