% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflated.R
\name{dist_inflated}
\alias{dist_inflated}
\title{Inflate a value of a probability distribution}
\usage{
dist_inflated(dist, prob, x = 0)
}
\arguments{
\item{dist}{The distribution(s) to inflate.}

\item{prob}{The added probability of observing \code{x}.}

\item{x}{The value to inflate. The default of \code{x = 0} is for zero-inflation.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
