% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_inverse_gamma.R
\name{dist_inverse_gamma}
\alias{dist_inverse_gamma}
\title{The Inverse Gamma distribution}
\usage{
dist_inverse_gamma(shape, rate = 1/scale, scale)
}
\arguments{
\item{shape}{parameters. Must be strictly positive.}

\item{rate}{an alternative way to specify the scale.}

\item{scale}{parameters. Must be strictly positive.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_inverse_gamma(shape = c(1,2,3,3), rate = c(1,1,1,2))

}
\seealso{
\link[actuar:InverseGamma]{actuar::InverseGamma}
}
