% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_inverse_gaussian.R
\name{dist_inverse_gaussian}
\alias{dist_inverse_gaussian}
\title{The Inverse Gaussian distribution}
\usage{
dist_inverse_gaussian(mean, shape)
}
\arguments{
\item{mean}{parameters. Must be strictly positive. Infinite
    values are supported.}

\item{shape}{parameters. Must be strictly positive. Infinite
    values are supported.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_inverse_gaussian(mean = c(1,1,1,3,3), shape = c(0.2, 1, 3, 0.2, 1))

}
\seealso{
\link[actuar:InverseGaussian]{actuar::InverseGaussian}
}
