% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_logarithmic.R
\name{dist_logarithmic}
\alias{dist_logarithmic}
\title{The Logarithmic distribution}
\usage{
dist_logarithmic(prob)
}
\arguments{
\item{prob}{parameter. \code{0 <= prob < 1}.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_logarithmic(prob = c(0.33, 0.66, 0.99))

}
\seealso{
\link[actuar:Logarithmic]{actuar::Logarithmic}
}
