% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_missing.R
\name{dist_missing}
\alias{dist_missing}
\title{Missing distribution}
\usage{
dist_missing(length = 1)
}
\arguments{
\item{length}{The number of missing distributions}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

A placeholder distribution for handling missing values in a vector of
distributions.
}
\examples{
dist <- dist_missing(3L)

dist
mean(dist)
variance(dist)

generate(dist, 10)

density(dist, 2)
density(dist, 2, log = TRUE)

cdf(dist, 4)

quantile(dist, 0.7)

}
