% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{dist_mixture}
\alias{dist_mixture}
\title{Create a mixture of distributions}
\usage{
dist_mixture(..., weights = numeric(), type = c("probability", "quantile"))
}
\arguments{
\item{...}{Distributions to be used in the mixture.}

\item{weights}{The weight of each distribution passed to \code{...}.}

\item{type}{The type of mixture to use, either 'probability' or 'quantile'.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\examples{
dist_mixture(dist_normal(0, 1), dist_normal(5, 2), weights = c(0.3, 0.7))

dist_df <- data.frame(
  type = c("N(1,3)", "N(10,1)", "probability", "quantile"),
  dist = c(
    dist_normal(1, 3),
    dist_normal(10, 1),
    dist_mixture(dist_normal(1, 3), dist_normal(10, 1), weights = c(0.5, 0.5), type = "probability"),
    dist_mixture(dist_normal(1, 3), dist_normal(10, 1), weights = c(0.5, 0.5), type = "quantile")
  )
)
library(ggdist)
library(ggplot2)
dist_df |>
  ggplot(aes(y = type, xdist = dist)) +
  stat_slab()

}
