% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{dist_mixture}
\alias{dist_mixture}
\title{Create a mixture of distributions}
\usage{
dist_mixture(..., weights = numeric())
}
\arguments{
\item{...}{Distributions to be used in the mixture.}

\item{weights}{The weight of each distribution passed to \code{...}.}
}
\description{
\lifecycle{experimental}
}
\examples{
dist_mixture(dist_normal(0, 1), dist_normal(5, 2), weights = c(0.3, 0.7))

}
