% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_multivariate_normal.R
\name{dist_multivariate_normal}
\alias{dist_multivariate_normal}
\title{The multivariate normal distribution}
\usage{
dist_multivariate_normal(mu = 0, sigma = diag(1))
}
\arguments{
\item{mu}{A list of numeric vectors for the distribution's mean.}

\item{sigma}{A list of matrices for the distribution's variance-covariance matrix.}
}
\description{
\lifecycle{maturing}
}
\examples{
dist_multivariate_normal(mu = list(c(1,2)), sigma = list(matrix(c(4,2,2,3), ncol=2)))

}
