% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_negative_binomial.R
\name{dist_negative_binomial}
\alias{dist_negative_binomial}
\title{The Negative Binomial distribution}
\usage{
dist_negative_binomial(size, prob)
}
\arguments{
\item{size}{target for number of successful trials, or dispersion
    parameter (the shape parameter of the gamma mixing distribution).
    Must be strictly positive, need not be integer.}

\item{prob}{probability of success in each trial. \code{0 < prob <= 1}.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_negative_binomial(size = 10, prob = 0.5)

}
\seealso{
\link[stats:NegBinomial]{stats::NegBinomial}
}
