% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_normal.R
\name{dist_normal}
\alias{dist_normal}
\title{The Normal distribution}
\usage{
dist_normal(mu = 0, sigma = 1)
}
\arguments{
\item{mu}{The mean (location parameter) of the distribution.}

\item{sigma}{The standard deviation (scale parameter) of the distribution.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_normal(mu = 1:5, sigma = 3)

}
\seealso{
\link[stats:Normal]{stats::Normal}
}
