% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_studentized_range.R
\name{dist_studentized_range}
\alias{dist_studentized_range}
\title{The Studentized Range distribution}
\usage{
dist_studentized_range(nmeans, df, nranges)
}
\arguments{
\item{nmeans}{sample size for range (same for each group).}

\item{df}{degrees of freedom for \eqn{s} (see below).}

\item{nranges}{number of \emph{groups} whose \bold{maximum} range is
    considered.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_studentized_range(nmeans = c(6, 2), df = c(5, 4), nranges = c(1, 1))

}
\seealso{
\link[stats:Tukey]{stats::Tukey}
}
