% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformed.R
\name{dist_transformed}
\alias{dist_transformed}
\title{Modify a distribution with a transformation}
\usage{
dist_transformed(dist, transform, inverse)
}
\arguments{
\item{dist}{A distribution vector}

\item{transform}{A function used to transform the distribution}

\item{inverse}{The inverse of the \code{transform} function}
}
\description{
Modify a distribution with a transformation
}
