% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_uniform.R
\name{dist_uniform}
\alias{dist_uniform}
\title{The Uniform distribution}
\usage{
dist_uniform(min, max)
}
\arguments{
\item{min}{lower and upper limits of the distribution.  Must be finite.}

\item{max}{lower and upper limits of the distribution.  Must be finite.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_uniform(min = c(3, -2), max = c(5, 4))

}
\seealso{
\link[stats:Uniform]{stats::Uniform}
}
