% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_weibull.R
\name{dist_weibull}
\alias{dist_weibull}
\title{The Weibull distribution}
\usage{
dist_weibull(shape, scale)
}
\arguments{
\item{shape}{shape and scale parameters, the latter defaulting to 1.}

\item{scale}{shape and scale parameters, the latter defaulting to 1.}
}
\description{
\lifecycle{stable}
}
\examples{
dist_weibull(shape = c(0.5, 1, 1.5, 5), scale = rep(1, 4))

}
\seealso{
\link[stats:Weibull]{stats::Weibull}
}
