% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributional-package.R
\docType{package}
\name{distributional-package}
\alias{distributional}
\alias{distributional-package}
\title{distributional: Vectorised Distributions}
\description{
Allows distributions to be used alongside data using vectors.
}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com} (\href{https://orcid.org/0000-0001-6729-7695}{ORCID})

}
\keyword{internal}
