% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributional-package.R
\docType{package}
\name{distributional-package}
\alias{distributional}
\alias{distributional-package}
\title{distributional: Vectorised Probability Distributions}
\description{
Vectorised distribution objects with tools for manipulating, 
    visualising, and using probability distributions. Designed to allow model
    prediction outputs to return distributions rather than their parameters, 
    allowing users to directly interact with predictive distributions in a
    data-oriented workflow. In addition to providing generic replacements for
    p/d/q/r functions, other useful statistics can be computed including means,
    variances, intervals, and highest density regions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.mitchelloharawild.com/distributional/}
  \item \url{https://github.com/mitchelloharawild/distributional}
  \item Report bugs at \url{https://github.com/mitchelloharawild/distributional/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com} (\href{https://orcid.org/0000-0001-6729-7695}{ORCID})

Other contributors:
\itemize{
  \item Earo Wang (\href{https://orcid.org/0000-0001-6448-5260}{ORCID}) [contributor]
}

}
\keyword{internal}
