% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{generate.distribution}
\alias{generate.distribution}
\title{Randomly sample values from a distribution}
\usage{
\method{generate}{distribution}(x, times, ...)
}
\arguments{
\item{x}{The distribution(s).}

\item{times}{The number of samples.}

\item{...}{Additional arguments used by methods.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Generate random samples from probability distributions.
}
