% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{is_distribution}
\alias{is_distribution}
\title{Test if the object is a distribution}
\usage{
is_distribution(x)
}
\arguments{
\item{x}{An object.}
}
\value{
TRUE if the object inherits from the distribution class.
}
\description{
This function returns \code{TRUE} for distributions and \code{FALSE} for all other objects.
\lifecycle{stable}
}
\examples{
dist <- dist_normal()
is_distribution(dist)
is_distribution("distributional")
}
