% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{mean.distribution}
\alias{mean.distribution}
\title{Mean of a probability distribution}
\usage{
\method{mean}{distribution}(x, ...)
}
\arguments{
\item{x}{The distribution(s).}

\item{...}{Additional arguments used by methods.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Returns the empirical mean of the probability distribution. If the method
does not exist, the mean of a random sample will be returned.
}
