% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{median.distribution}
\alias{median.distribution}
\title{Median of a probability distribution}
\usage{
\method{median}{distribution}(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{The distribution(s).}

\item{na.rm}{Unused, included for consistency with the generic function.}

\item{...}{Additional arguments used by methods.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Returns the median (50th percentile) of a probability distribution. This is
equivalent to \code{quantile(x, p=0.5)}.
}
