% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdr.R
\name{new_hdr}
\alias{new_hdr}
\title{Construct hdr intervals}
\usage{
new_hdr(x = list())
}
\arguments{
\item{x}{A list of \code{\link[=hilo]{hilo()}} objects.}
}
\value{
A "hdr" vector
}
\description{
Construct hdr intervals
}
\author{
Mitchell O'Hara-Wild
}
