% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilo.R
\name{new_hilo}
\alias{new_hilo}
\title{Construct hilo intervals}
\usage{
new_hilo(lower = double(), upper = double(), size = double())
}
\arguments{
\item{lower, upper}{A numeric vector of values for lower and upper limits.}

\item{size}{Size of the interval between [0, 100].}
}
\value{
A "hilo" vector
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Class constructor function to help with manually creating hilo interval
objects.
}
\examples{
new_hilo(lower = rnorm(10), upper = rnorm(10) + 5, size = 95)

}
\author{
Earo Wang & Mitchell O'Hara-Wild
}
