% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{new_support_region}
\alias{new_support_region}
\title{Create a new support region vector}
\usage{
new_support_region(x = numeric(), limits = list(), closed = list())
}
\arguments{
\item{x}{A list of prototype vectors defining the distribution type.}

\item{limits}{A list of value limits for the distribution.}

\item{closed}{A list of logical(2L) indicating whether the limits are closed.}
}
\description{
Create a new support region vector
}
