% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{parameters}
\alias{parameters}
\title{Extract the parameters of a distribution}
\usage{
parameters(x, ...)
}
\arguments{
\item{x}{The distribution(s).}

\item{...}{Additional arguments used by methods.}
}
\description{
\lifecycle{experimental}
}
