% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{parameters}
\alias{parameters}
\alias{parameters.distribution}
\title{Extract the parameters of a distribution}
\usage{
parameters(x, ...)

\method{parameters}{distribution}(x, ...)
}
\arguments{
\item{x}{The distribution(s).}

\item{...}{Additional arguments used by methods.}
}
\description{
\lifecycle{experimental}
}
\examples{
dist <- c(
  dist_normal(1:2),
  dist_poisson(3),
  dist_multinomial(size = c(4, 3),
  prob = list(c(0.3, 0.5, 0.2), c(0.1, 0.5, 0.4)))
  )
parameters(dist)
}
