% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{quantile.distribution}
\alias{quantile.distribution}
\title{Distribution Quantiles}
\usage{
\method{quantile}{distribution}(x, p, ...)
}
\arguments{
\item{x}{The distribution(s).}

\item{p}{The probability of the quantile.}

\item{...}{Additional arguments passed to methods.}
}
\description{
\lifecycle{stable}
}
\details{
Computes the quantiles of a distribution.
}
