% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{support}
\alias{support}
\alias{support.distribution}
\title{Region of support of a distribution}
\usage{
support(x, ...)

\method{support}{distribution}(x, ...)
}
\arguments{
\item{x}{The distribution(s).}

\item{...}{Additional arguments used by methods.}
}
\description{
\lifecycle{experimental}
}
