% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{variance}
\alias{variance}
\title{Variance}
\usage{
variance(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments used by methods.}
}
\description{
A generic function for computing the variance of an object. The default
method will use \code{\link[stats:var]{stats::var()}} to compute the variance.
}
\seealso{
\code{\link[=variance.distribution]{variance.distribution()}}
}
